Move data table name normalisation to utility functions
