+++ src/config.ts
@@ -34,7 +34,11 @@ const config: { [key: string]: any } = {
     process.env.NODE_ENV === 'production' ? '../../node_modules' : '../node_modules',
   jwtSecret: process.env.JWT_SECRET || 'devsecret',
   RESTport: 8080,
-  dataTablePrefix: 'data_table',
+  dataTablePrefix: 'data_table_',
+  // These are the only default tables in the system that we allow to be mutated
+  // directly by modifyRecord or display as data views. All other names must
+  // have "data_table_" prepended.
+  allowedTableNames: ['user', 'organisation', 'application', 'file'],
   ...serverPrefs,
 }
